Velocity/backend/requirements.txt
Flask==3.0.0
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
pydantic==2.5.3
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.6.1
bcrypt==4.1.2
pytest==7.4.4
pytest-asyncio==0.23.2
httpx==0.26.0